refactor(web): ROK-1082 share TasteRadarChart between player + game pages
Moves TasteRadarChart, ArchetypePill, taste-profile-helpers, and
taste-profile-section.css from pages/user-profile/taste-profile/ to
shared components/taste-profile/. Makes archetype prop optional so the
game-detail page can use the same chart (no archetype for games). Adds
a dataTestId prop so existing game-radar-chart selectors still work.

Deletes GameRadarChart.tsx — the 20-line parallel is no longer needed.

Consumer import updates:
- TasteProfileSection.tsx (user profile)
- user-profile-page.tsx
- IntensityBadge.tsx
- AxisBreakdown.tsx (game side)
- GameTasteSection.tsx now renders TasteRadarChart with no archetype

All 2897 web tests pass.
