+++ web/src/components/taste-profile/ArchetypePill.tsx
