+++ web/src/components/taste-profile/taste-profile-helpers.ts
