+++ web/src/components/taste-profile/taste-profile-section.css
